Data.Diff

Documentation

class (Eq a, Patch (Edit a)) => Diff a where

Diffable types

uncurry diff . undiff = id
undiff . uncurry diff = id
patch (diff x y) x = Just y

Associated Types

type Edit a

Methods

diff :: a -> a -> Edit a

Generate a patch between two values, that can be used to transform one value into the other

patch :: Edit a -> a -> Maybe a

Apply a patch to a value

undiff :: Edit a -> (a, a)

Deconstruct a patch to recover the original value used to create it, and the application of the patch to that original value

diff :: DefaultDiff (Edit a) a => a -> a -> Edit a

Generate a patch between two values, that can be used to transform one value into the other

patch :: DefaultDiff (Edit a) a => Edit a -> a -> Maybe a

Apply a patch to a value

undiff :: DefaultDiff (Edit a) a => Edit a -> (a, a)

Deconstruct a patch to recover the original value used to create it, and the application of the patch to that original value

class Patch a where

Methods

patchLevel :: a -> DiffLevel

Level of patch

mergePatch :: a -> a -> MergeResult a

Left-biased parallel merge of two patches

Returns Nothing if patches come from incompatible sources

Returns True if conflict occurred (and was resolved)

patchLevel :: (Generic a, Every (Every Patch) (Code a)) => a -> DiffLevel

Level of patch

mergePatch :: (IsProductType a as, Every Patch as) => a -> a -> MergeResult a

Left-biased parallel merge of two patches

Returns Nothing if patches come from incompatible sources

Returns True if conflict occurred (and was resolved)

data DiffLevel

Data type representing a "percentage difference"

Constructors

DL

Fields

dlAmt :: Double

dlTot :: Double

Bundled Patterns

pattern NoDiff :: Double -> DiffLevel

pattern TotalDiff :: Double -> DiffLevel

data MergeResult a

Result of a merge

Constructors

Incompatible
Incompatible sources

Conflict a
Conflicts, throwing away info

NoConflict a
All conflicts resolved

merge :: Diff a => a -> a -> a -> MergeResult a

Left-biased merge of two diffable values.

catLevels :: Foldable f => f DiffLevel -> DiffLevel

Merge all DiffLevel in a Foldable container, treating them all as equally weighted.

normDL :: Double -> DiffLevel -> DiffLevel

Rescale DiffLevel to be out of a given total.

TODO: what about 0?

dlPercent :: DiffLevel -> Double

Calculate percent difference.

TODO: What about 0?

percentDiff :: Diff a => a -> a -> Double

How different two items are, as a percentage

prodPatchLevel :: forall as. Every Diff as => Prod Edit' as -> DiffLevel

DiffLevel of a Prod of Edits

compareDiff :: Diff a => a -> a -> DiffLevel

Get DiffLevel between two items

class DefaultDiff p a where

Minimal complete definition

defaultDiff, defaultPatch, defaultUndiff

Methods

defaultDiff :: a -> a -> p

defaultPatch :: p -> a -> Maybe a

defaultUndiff :: p -> (a, a)

newtype Edit' a

Newtype used to get around partial application of type families

Constructors

Edit'

Fields

getEdit' :: Edit a

diff' :: Diff a => a -> a -> Edit' a

diff' lifted to Edit'

patch' :: Diff a => Edit' a -> a -> Maybe a

patch' lifted to Edit'

undiff' :: Diff a => Edit' a -> (a, a)

undiff' lifted to Edit'

data Swap a

Patch type that treats all changes as total differences

Constructors

NoChange a

Replace a a

eqDiff :: Eq a => a -> a -> Swap a

diff for all instances of Eq

eqPatch :: Swap a -> a -> Maybe a

patch for Swap

newtype EqDiff a

Newtype wrapper that gives an automatic Diff instance that treats all changes as total differences.

Constructors

EqDiff

Fields

getEqDiff :: a

Generic implementations

Patch

gpatchLevel :: forall a ass. (Generic a, Code a ~ ass, Every (Every Patch) ass) => a -> DiffLevel

patchLevel written to work for any types deriving Generic.

gmergePatch :: forall a as. (IsProductType a as, Every Patch as) => a -> a -> MergeResult a

mergePatch written to work for any product types deriving Generic.

Diff

Sequences

listDiff :: forall p a. (KnownNat p, Diff a) => [a] -> [a] -> SeqPatchAt p a

listPatch :: Eq a => SeqPatchAt p a -> [a] -> Maybe [a]

newtype SeqPatchAt (p :: Nat) a

Constructors

SPA

Fields

getSPA :: [Diff a]

type SeqPatch = SeqPatchAt 50

seqDiff :: (KnownNat p, Diff a) => (t -> [a]) -> t -> t -> SeqPatchAt p a

seqPatch :: Eq a => (t -> [a]) -> ([a] -> t) -> SeqPatchAt p a -> t -> Maybe t

Generic ADTs

newtype GPatch a

Generic patch type for any types deriving Generic.

Constructors

GP

Fields

getGP :: SumDiff Tuple (Prod Edit') (Code a)

gdiff :: forall a. (Generic a, Every (Every Diff) (Code a), Every Typeable (Code a)) => a -> a -> GPatch a

diff intented to work for all instances of Generic. Will treat constructor changes as partial differences if they both contain the same type of values.

gdiff' :: forall a. (Generic a, Every (Every Diff) (Code a)) => a -> a -> GPatch a

diff intented to work for all instances of Generic. Differs from gdiff in that it treats constructor changes as total differences, even if they both contain the same type of values.

gpatch :: forall a. (Generic a, Every (Every Diff) (Code a)) => GPatch a -> a -> Maybe a

patch intented to work for all instances of Generic.

gundiff :: forall a. (Generic a, Every (Every Diff) (Code a)) => GPatch a -> (a, a)

undiff intended to work for all instances of Generic.

data GPatchProd a

Generic patch type for all product types that are instances of Generic.

Constructors

Code a ~ '[as] => GPP

Fields

getGPP :: Prod Edit' as

gdiffProd :: forall a as. (IsProductType a as, Every Diff as) => a -> a -> GPatchProd a

diff intended to work for all product types that are instances of Generic.

gpatchProd :: forall a as. (IsProductType a as, Every Diff as) => GPatchProd a -> a -> Maybe a

patch intended to work for all product types that are instances of Generic.

gundiffProd :: forall a as. (IsProductType a as, Every Diff as) => GPatchProd a -> (a, a)

undiff intended to work for all product types that are instances of Generic.

data SumDiff :: (k -> Type) -> (k -> Type) -> [k] -> Type where

Constructors

SD :: (Index as :&: CtrDiff f g as) a -> SumDiff f g as

data CtrDiff :: (k -> Type) -> (k -> Type) -> [k] -> k -> Type where

Constructors

CDEdit :: g a -> CtrDiff f g as a

CDName :: (Index as :&: g) a -> CtrDiff f g as a

CDDiff :: f a -> (Index as :&: f) b -> CtrDiff f g as a

Maps

data ValDiff a

Constructors

VDDel a

VDIns a

VDMod (Edit a)

newtype MapDiff m a

Constructors

MD

Fields

getMD :: m (ValDiff a)